Move calculateAmountReceived to SdkBase
