+++ packages/agents/sdk/src/sdkPool.ts
@@ -5,7 +5,6 @@ import {
   Logger,
   createLoggingContext,
   ChainData,
-  DEFAULT_ROUTER_FEE,
   formatUrl,
   StableSwapExchange,
 } from "@connext/nxtp-utils";
@@ -317,93 +316,6 @@ export class SdkPool extends SdkShared {
       BigNumber.from(10).pow(18),
       false,
     );
-  }
-
-  /**
-   * Calculates the estimated amount received on the destination domain for a bridge transaction.
-   *
-   * @param originDomain - The domain ID of the origin chain.
-   * @param destinationDomain - The domain ID of the destination chain.
-   * @param originTokenAddress - The address of the token to be bridged from origin.
-   * @param amount - The amount of the origin token to bridge, in the origin token's native decimal precision.
-   * @param receiveLocal - (optional) Whether the desired destination token is the local asset ("nextAsset").
-   * @returns Estimated amount received for local/adopted assets, if applicable, in their native decimal precisions.
-   */
-  async calculateAmountReceived(
-    originDomain: string,
-    destinationDomain: string,
-    originTokenAddress: string,
-    amount: BigNumberish,
-    receiveLocal = false,
-  ): Promise<{
-    amountReceived: BigNumberish;
-    originSlippage: BigNumberish;
-    routerFee: BigNumberish;
-    destinationSlippage: BigNumberish;
-  }> {
-    const { requestContext, methodContext } = createLoggingContext(this.calculateAmountReceived.name);
-
-    const _originTokenAddress = utils.getAddress(originTokenAddress);
-
-    this.logger.info("Method start", requestContext, methodContext, {
-      originDomain,
-      destinationDomain,
-      _originTokenAddress,
-      amount,
-    });
-
-    // Calculate origin swap
-    const originPool = await this.getPool(originDomain, _originTokenAddress);
-    let originAmountReceived = amount;
-
-    // Swap IFF supplied origin token is an adopted asset
-    if (!(await this.isNextAsset(_originTokenAddress)) && originPool) {
-      originAmountReceived = await this.calculateSwap(
-        originDomain,
-        _originTokenAddress,
-        originPool.adopted.index,
-        originPool.local.index,
-        amount,
-      );
-    }
-
-    const originSlippage = BigNumber.from(amount).sub(originAmountReceived).mul(10000).div(amount);
-    const feeBps = BigNumber.from(+DEFAULT_ROUTER_FEE * 100);
-    const routerFee = BigNumber.from(originAmountReceived).mul(feeBps).div(10000);
-
-    // Calculate destination swap
-    const [canonicalDomain, canonicalId] = await this.getCanonicalTokenId(originDomain, _originTokenAddress);
-    const key = this.calculateCanonicalKey(canonicalDomain, canonicalId);
-    const destinationAssetData = await this.getAssetsDataByDomainAndKey(destinationDomain, key);
-    if (!destinationAssetData) {
-      throw new Error("Origin token cannot be bridged to any token on this destination domain");
-    }
-
-    const destinationPool = await this.getPool(destinationDomain, destinationAssetData.local);
-    const destinationAmount = BigNumber.from(originAmountReceived).sub(routerFee);
-    let destinationAmountReceived = destinationAmount;
-
-    // Swap IFF desired destination token is an adopted asset
-    if (!receiveLocal && destinationPool) {
-      destinationAmountReceived = await this.calculateSwap(
-        destinationDomain,
-        destinationAssetData.local,
-        destinationPool.local.index,
-        destinationPool.adopted.index,
-        destinationAmount,
-      );
-    }
-
-    const destinationSlippage = BigNumber.from(
-      destinationAmount.sub(destinationAmountReceived).mul(10000).div(destinationAmount),
-    );
-
-    return {
-      amountReceived: destinationAmountReceived,
-      originSlippage,
-      routerFee,
-      destinationSlippage,
-    };
   }
 
   /**
