+++ packages/agents/sdk/test/sdkBase.spec.ts
@@ -1,10 +1,19 @@
 import { reset, restore, stub, SinonStub, createStubInstance, SinonStubbedInstance } from "sinon";
-import { encodeMultisendCall, expect, MultisendTransaction, WETHAbi, mkAddress } from "@connext/nxtp-utils";
+import {
+  encodeMultisendCall,
+  expect,
+  MultisendTransaction,
+  WETHAbi,
+  mkAddress,
+  DEFAULT_ROUTER_FEE,
+} from "@connext/nxtp-utils";
 import { getConnextInterface, ChainReader } from "@connext/nxtp-txservice";
-import { providers, BigNumber, utils } from "ethers";
+import { providers, BigNumber, utils, constants } from "ethers";
 import { mock } from "./mock";
 import { SdkBase } from "../src/sdkBase";
 import { SdkUtils } from "../src/sdkUtils";
+import { SdkPool } from "../src/sdkPool";
+import { PoolAsset, Pool } from "../src/interfaces";
 import { getEnvConfig } from "../src/config";
 import { CannotUnwrapOnDestination, ParamsInvalid, SignerAddressMissing } from "../src/lib/errors";
 
@@ -24,6 +33,7 @@ const chainId = +mock.chain.A;
 describe("SdkBase", () => {
   let sdkBase: SdkBase;
   let sdkUtils: SdkUtils;
+  let sdkPool: SdkPool;
   let config: ConfigFns.SdkConfig;
 
   let chainreader: SinonStubbedInstance<ChainReader>;
@@ -37,6 +47,7 @@ describe("SdkBase", () => {
 
     sdkBase = await SdkBase.create(mockConfig, undefined, mockChainData);
     sdkUtils = await SdkUtils.create(mockConfig, undefined, mockChainData);
+    sdkPool = await SdkPool.create(mockConfig, undefined, mockChainData);
     (sdkBase as any).chainreader = chainreader;
   });
 
@@ -55,6 +66,7 @@ describe("SdkBase", () => {
       expect(sdkBase.bumpTransfer).to.be.a("function");
       expect(sdkBase.updateSlippage).to.be.a("function");
       expect(sdkBase.estimateRelayerFee).to.be.a("function");
+      expect(sdkBase.calculateAmountReceived).to.be.a("function");
     });
   });
 
@@ -427,6 +439,155 @@ describe("SdkBase", () => {
       });
       expect(calculateRelayerFeeStub.callCount).to.be.eq(1);
       expect(relayerFee.toString()).to.be.eq("100");
+    });
+  });
+
+  describe("#calculateAmountReceived", () => {
+    const localAsset: PoolAsset = {
+      address: mock.asset.A.address,
+      name: mock.asset.A.name,
+      symbol: mock.asset.A.symbol,
+      decimals: 18,
+      index: 0,
+      balance: BigNumber.from("100"),
+    };
+
+    const adoptedAsset: PoolAsset = {
+      address: mock.asset.B.address,
+      name: mock.asset.B.name,
+      symbol: mock.asset.B.symbol,
+      decimals: 18,
+      index: 1,
+      balance: BigNumber.from("100"),
+    };
+
+    const mockPool: Pool = {
+      domainId: mock.domain.A,
+      name: "TSTB Pool",
+      symbol: "TSTB-TSTA",
+      local: localAsset,
+      adopted: adoptedAsset,
+      lpTokenAddress: utils.formatBytes32String("asdf"),
+      canonicalHash: utils.formatBytes32String("13337"),
+      swapFee: "4000000",
+      adminFee: "0",
+    };
+
+    const mockAssetData = {
+      local: mockPool.local.address,
+      adopted: mockPool.adopted.address,
+      canonical_id: constants.HashZero,
+      canonical_domain: mockPool.domainId,
+      domain: mockPool.domainId,
+      key: mockPool.canonicalHash,
+      id: mockPool.local.address,
+    };
+
+    const feeBps = BigNumber.from(+DEFAULT_ROUTER_FEE * 100);
+
+    it("happy: should work with local origin asset and adopted destination asset", async () => {
+      stub(sdkPool, "getPool").onCall(0).resolves(undefined).onCall(1).resolves(mockPool);
+
+      const originAmount = BigNumber.from(100_000);
+      const originSlippage = "0"; // 0% in BPS
+
+      const destinationAmount = originAmount.sub(originAmount.mul(feeBps).div(10000)); // router takes 0.05%
+      const destinationAmountAfterSwap = destinationAmount.mul(9).div(10); // assume swap ate 10%;
+      const destinationSlippage = "1000"; // 10% in BPS
+
+      stub(sdkPool, "calculateSwap")
+        .onCall(0) // swap once for destination pool
+        .resolves(destinationAmountAfterSwap);
+      stub(sdkBase, "getCanonicalTokenId").resolves([mockAssetData.canonical_domain, mockAssetData.canonical_id]);
+      stub(sdkBase, "getAssetsDataByDomainAndKey").resolves(mockAssetData);
+
+      const res = await sdkBase.calculateAmountReceived(
+        mockPool.domainId,
+        mockPool.domainId,
+        mockPool.local.address,
+        originAmount,
+      );
+
+      expect(res.originSlippage.toString()).to.equal(originSlippage);
+      expect(res.destinationSlippage.toString()).to.equal(destinationSlippage);
+    });
+
+    it("happy: should work with adopted origin asset and adopted destination asset", async () => {
+      stub(sdkPool, "getPool").onCall(0).resolves(mockPool).onCall(1).resolves(mockPool);
+
+      const originAmount = BigNumber.from(100_000);
+      const originAmountAfterSwap = originAmount.mul(9).div(10); // assume swap ate 10%
+      const originSlippage = "1000"; // 10% in BPS
+
+      const destinationAmount = originAmountAfterSwap.sub(originAmountAfterSwap.mul(feeBps).div(10000)); // router takes 0.05%
+      const destinationAmountAfterSwap = destinationAmount.mul(9).div(10); // assume swap ate 10%;
+      const destinationSlippage = "1000"; // 10% in BPS
+
+      stub(sdkPool, "calculateSwap")
+        .onCall(0) // swap once for origin pool
+        .resolves(originAmountAfterSwap)
+        .onCall(1) // swap once for destination pool
+        .resolves(destinationAmountAfterSwap);
+      stub(sdkBase, "getCanonicalTokenId").resolves([mockAssetData.canonical_domain, mockAssetData.canonical_id]);
+      stub(sdkBase, "getAssetsDataByDomainAndKey").resolves(mockAssetData);
+
+      const res = await sdkBase.calculateAmountReceived(
+        mockPool.domainId,
+        mockPool.domainId,
+        mockPool.adopted.address,
+        originAmount,
+      );
+
+      expect(res.originSlippage.toString()).to.equal(originSlippage);
+      expect(res.destinationSlippage.toString()).to.equal(destinationSlippage);
+    });
+
+    it("happy: should work with adopted origin asset and local destination asset", async () => {
+      stub(sdkPool, "getPool").onCall(0).resolves(mockPool).onCall(1).resolves(undefined);
+
+      const originAmount = BigNumber.from(100_000);
+      const originAmountAfterSwap = originAmount.mul(9).div(10); // assume swap ate 10%
+      const originSlippage = "1000"; // 10% in BPS
+      const destinationSlippage = "0"; // 0% in BPS
+
+      stub(sdkPool, "calculateSwap")
+        .onCall(0) // swap once for origin pool
+        .resolves(originAmountAfterSwap);
+      stub(sdkBase, "getCanonicalTokenId").resolves([mockAssetData.canonical_domain, mockAssetData.canonical_id]);
+      stub(sdkBase, "getAssetsDataByDomainAndKey").resolves(mockAssetData);
+
+      const res = await sdkBase.calculateAmountReceived(
+        mockPool.domainId,
+        mockPool.domainId,
+        mockPool.adopted.address,
+        originAmount,
+        true,
+      );
+
+      expect(res.originSlippage.toString()).to.equal(originSlippage);
+      expect(res.destinationSlippage.toString()).to.equal(destinationSlippage);
+    });
+
+    it("happy: should work with local origin asset and local destination asset", async () => {
+      stub(sdkPool, "getPool").onCall(0).resolves(undefined).onCall(1).resolves(undefined);
+
+      const originAmount = BigNumber.from(100_000);
+      const originSlippage = "0"; // 10% in BPS
+      const destinationSlippage = "0"; // 0% in BPS
+
+      stub(sdkBase, "getCanonicalTokenId").resolves([mockAssetData.canonical_domain, mockAssetData.canonical_id]);
+      stub(sdkBase, "getAssetsDataByDomainAndKey").resolves(mockAssetData);
+
+      const res = await sdkBase.calculateAmountReceived(
+        mockPool.domainId,
+        mockPool.domainId,
+        mockPool.local.address,
+        originAmount,
+        true,
+      );
+
+      expect(res.originSlippage.toString()).to.equal(originSlippage);
+      expect(res.destinationSlippage.toString()).to.equal(destinationSlippage);
     });
   });
 });
