+++ packages/agents/sdk/test/sdkPool.spec.ts
@@ -1,12 +1,11 @@
-import { reset, restore, stub, spy } from "sinon";
+import { reset, restore, stub } from "sinon";
 import { expect, getCanonicalHash, DEFAULT_ROUTER_FEE, getRandomBytes32 } from "@connext/nxtp-utils";
 import { getConnextInterface } from "@connext/nxtp-txservice";
-import { providers, utils, BigNumber, Contract, constants } from "ethers";
+import { providers, utils, BigNumber, constants } from "ethers";
 import { mock } from "./mock";
 import { SdkPool } from "../src/sdkPool";
 import { PoolAsset, Pool } from "../src/interfaces";
 import { getEnvConfig } from "../src/config";
-import { Connext } from "@connext/smart-contracts";
 
 import * as ConfigFns from "../src/config";
 import { UriInvalid } from "../src/lib/errors";
@@ -72,7 +71,7 @@ describe("SdkPool", () => {
   const mockAssetData = {
     local: mockPool.local.address,
     adopted: mockPool.adopted.address,
-    canonical_id: utils.formatBytes32String("0"),
+    canonical_id: constants.HashZero,
     canonical_domain: mockPool.domainId,
     domain: mockPool.domainId,
     key: mockPool.canonicalHash,
@@ -109,7 +108,6 @@ describe("SdkPool", () => {
       expect(sdkPool.calculateSwapPriceImpact).to.be.a("function");
       expect(sdkPool.calculateAddLiquidityPriceImpact).to.be.a("function");
       expect(sdkPool.calculateRemoveLiquidityPriceImpact).to.be.a("function");
-      expect(sdkPool.calculateAmountReceived).to.be.a("function");
       expect(sdkPool.getTokenPrice).to.be.a("function");
       expect(sdkPool.getDefaultDeadline).to.be.a("function");
 
@@ -410,125 +408,6 @@ describe("SdkPool", () => {
       const res = await sdkPool.calculatePriceImpact(BigNumber.from(0), BigNumber.from(0), mockParams.virtualPrice);
 
       expect(res.toString()).to.equal(BigNumber.from("0").toString());
-    });
-  });
-
-  describe("#calculateAmountReceived", () => {
-    const mockAssetData = {
-      local: mockPool.local.address,
-      adopted: mockPool.adopted.address,
-      canonical_id: utils.formatBytes32String("0"),
-      canonical_domain: mockPool.domainId,
-      domain: mockPool.domainId,
-      key: mockPool.canonicalHash,
-      id: mockPool.local.address,
-    };
-
-    const feeBps = BigNumber.from(+DEFAULT_ROUTER_FEE * 100);
-
-    it("happy: should work with local origin asset and adopted destination asset", async () => {
-      stub(sdkPool, "getPool").onCall(0).resolves(undefined).onCall(1).resolves(mockPool);
-
-      const originAmount = BigNumber.from(100_000);
-      const originSlippage = "0"; // 0% in BPS
-
-      const destinationAmount = originAmount.sub(originAmount.mul(feeBps).div(10000)); // router takes 0.05%
-      const destinationAmountAfterSwap = destinationAmount.mul(9).div(10); // assume swap ate 10%;
-      const destinationSlippage = "1000"; // 10% in BPS
-
-      stub(sdkPool, "calculateSwap")
-        .onCall(0) // swap once for destination pool
-        .resolves(destinationAmountAfterSwap);
-      stub(sdkPool, "getCanonicalTokenId").resolves([mockAssetData.canonical_domain, mockAssetData.canonical_id]);
-      stub(sdkPool, "getAssetsDataByDomainAndKey").resolves(mockAssetData);
-
-      const res = await sdkPool.calculateAmountReceived(
-        mockPool.domainId,
-        mockPool.domainId,
-        mockPool.local.address,
-        originAmount,
-      );
-
-      expect(res.originSlippage.toString()).to.equal(originSlippage);
-      expect(res.destinationSlippage.toString()).to.equal(destinationSlippage);
-    });
-
-    it("happy: should work with adopted origin asset and adopted destination asset", async () => {
-      stub(sdkPool, "getPool").onCall(0).resolves(mockPool).onCall(1).resolves(mockPool);
-
-      const originAmount = BigNumber.from(100_000);
-      const originAmountAfterSwap = originAmount.mul(9).div(10); // assume swap ate 10%
-      const originSlippage = "1000"; // 10% in BPS
-
-      const destinationAmount = originAmountAfterSwap.sub(originAmountAfterSwap.mul(feeBps).div(10000)); // router takes 0.05%
-      const destinationAmountAfterSwap = destinationAmount.mul(9).div(10); // assume swap ate 10%;
-      const destinationSlippage = "1000"; // 10% in BPS
-
-      stub(sdkPool, "calculateSwap")
-        .onCall(0) // swap once for origin pool
-        .resolves(originAmountAfterSwap)
-        .onCall(1) // swap once for destination pool
-        .resolves(destinationAmountAfterSwap);
-      stub(sdkPool, "getCanonicalTokenId").resolves([mockAssetData.canonical_domain, mockAssetData.canonical_id]);
-      stub(sdkPool, "getAssetsDataByDomainAndKey").resolves(mockAssetData);
-
-      const res = await sdkPool.calculateAmountReceived(
-        mockPool.domainId,
-        mockPool.domainId,
-        mockPool.adopted.address,
-        originAmount,
-      );
-
-      expect(res.originSlippage.toString()).to.equal(originSlippage);
-      expect(res.destinationSlippage.toString()).to.equal(destinationSlippage);
-    });
-
-    it("happy: should work with adopted origin asset and local destination asset", async () => {
-      stub(sdkPool, "getPool").onCall(0).resolves(mockPool).onCall(1).resolves(undefined);
-
-      const originAmount = BigNumber.from(100_000);
-      const originAmountAfterSwap = originAmount.mul(9).div(10); // assume swap ate 10%
-      const originSlippage = "1000"; // 10% in BPS
-      const destinationSlippage = "0"; // 0% in BPS
-
-      stub(sdkPool, "calculateSwap")
-        .onCall(0) // swap once for origin pool
-        .resolves(originAmountAfterSwap);
-      stub(sdkPool, "getCanonicalTokenId").resolves([mockAssetData.canonical_domain, mockAssetData.canonical_id]);
-      stub(sdkPool, "getAssetsDataByDomainAndKey").resolves(mockAssetData);
-
-      const res = await sdkPool.calculateAmountReceived(
-        mockPool.domainId,
-        mockPool.domainId,
-        mockPool.adopted.address,
-        originAmount,
-        true,
-      );
-
-      expect(res.originSlippage.toString()).to.equal(originSlippage);
-      expect(res.destinationSlippage.toString()).to.equal(destinationSlippage);
-    });
-
-    it("happy: should work with local origin asset and local destination asset", async () => {
-      stub(sdkPool, "getPool").onCall(0).resolves(undefined).onCall(1).resolves(undefined);
-
-      const originAmount = BigNumber.from(100_000);
-      const originSlippage = "0"; // 10% in BPS
-      const destinationSlippage = "0"; // 0% in BPS
-
-      stub(sdkPool, "getCanonicalTokenId").resolves([mockAssetData.canonical_domain, mockAssetData.canonical_id]);
-      stub(sdkPool, "getAssetsDataByDomainAndKey").resolves(mockAssetData);
-
-      const res = await sdkPool.calculateAmountReceived(
-        mockPool.domainId,
-        mockPool.domainId,
-        mockPool.local.address,
-        originAmount,
-        true,
-      );
-
-      expect(res.originSlippage.toString()).to.equal(originSlippage);
-      expect(res.destinationSlippage.toString()).to.equal(destinationSlippage);
     });
   });
